pyproject.toml
[tool.poetry]
name = "tfg-h16b"
version = "0.1.0"
description = ""
authors = ["Víctor González Prieto <victor.gonzalez.prieto@estudiantat.upc.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.6.9"
numpy = "1.16.3"
scipy = "1.2.1"
matplotlib = "=2.1.0"
PyYAML = "5.1"
numba = "0.37.0"

# sub-dependencies:
llvmlite = "0.22.0"
Cycler = "0.10.0"
pyparsing = "2.4.0"
python-dateutil = "2.6.1"
pytz = "2017.3"
six = "1.12.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
